refact: rename core.mastodon.feed.profile.toot.cache.storage.mention to 'style'
