+++ core/mastodon/src/main/java/com/jeanbarrossilva/orca/core/mastodon/feed/profile/toot/cache/storage/MastodonTootEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 import com.jeanbarrossilva.orca.core.feed.profile.Profile
 import com.jeanbarrossilva.orca.core.feed.profile.toot.style.toStyledString
 import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.MastodonToot
-import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.cache.storage.mention.startingAt
+import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.cache.storage.style.startingAt
 import com.jeanbarrossilva.orca.platform.cache.Cache
 import java.net.URL
 import java.time.ZonedDateTime
