+++ core/mastodon/src/main/java/com/jeanbarrossilva/orca/core/mastodon/feed/profile/toot/cache/storage/MastodonTootWithMentions.kt
@@ -2,7 +2,7 @@ package com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.cache.storage
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.cache.storage.mention.StyleEntity
+import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.cache.storage.style.StyleEntity
 
 internal data class MastodonTootWithMentions(
     @Embedded val toot: MastodonTootEntity,
